feat: webp to png on-the-fly
