+++ src/components/App.tsx
@@ -68,19 +68,19 @@ import { SettingContext } from "../context";
 import { ContentTransModeType, DisplayModeType } from "../types";
 
 const drawerWidth = 240;
-const CardList = lazy(() => import("./CardList"));
+const CardList = lazy(() => import("./card/CardList"));
 const HomeView = lazy(() => import("./Home"));
-const MusicList = lazy(() => import("./MusicList"));
-const GachaList = lazy(() => import("./GachaList"));
-const EventList = lazy(() => import("./EventList"));
-const GachaDetail = lazy(() => import("./GachaDetail"));
-const CardDetail = lazy(() => import("./CardDetail"));
-const MusicDetail = lazy(() => import("./MusicDetail"));
-const EventDetail = lazy(() => import("./EventDetail"));
-const ComicList = lazy(() => import("./ComicList"));
+const MusicList = lazy(() => import("./music/MusicList"));
+const GachaList = lazy(() => import("./gacha/GachaList"));
+const EventList = lazy(() => import("./event/EventList"));
+const GachaDetail = lazy(() => import("./gacha/GachaDetail"));
+const CardDetail = lazy(() => import("./card/CardDetail"));
+const MusicDetail = lazy(() => import("./music/MusicDetail"));
+const EventDetail = lazy(() => import("./event/EventDetail"));
+const ComicList = lazy(() => import("./comic/ComicList"));
 const MemberDetail = lazy(() => import("./MemberDetail"));
 const MemberList = lazy(() => import("./MemberList"));
-const StampList = lazy(() => import("./StampList"));
+const StampList = lazy(() => import("./stamp/StampList"));
 const UnitDetail = lazy(() => import("./UnitDetail"));
 const About = lazy(() => import("./About"));
 const Support = lazy(() => import("./Support"));
