+++ src/components/card/AgendaView.tsx
@@ -1,0 +1,137 @@
+import { Typography, Grid, makeStyles, Box, Paper } from "@material-ui/core";
+import { Skeleton } from "@material-ui/lab";
+import React, { useContext } from "react";
+import { Link, useRouteMatch } from "react-router-dom";
+import { SettingContext } from "../../context";
+import { ICardInfo } from "../../types";
+import { useCharaName } from "../../utils";
+import { CardThumbSkeleton, CardThumb } from "../subs/CardThumb";
+import { ContentTrans } from "../subs/ContentTrans";
+
+const useStyles = makeStyles((theme) => ({
+  media: {
+    paddingTop: "56.25%",
+  },
+  card: {
+    // margin: theme.spacing(0.5),
+    cursor: "pointer",
+  },
+  subheader: {
+    // whiteSpace: "nowrap",
+    // overflow: "hidden",
+    // textOverflow: "ellipsis",
+  },
+  agendaWrapper: {
+    display: "block",
+    [theme.breakpoints.down("sm")]: {
+      maxWidth: "100%",
+    },
+    // [theme.breakpoints.only("md")]: {
+    //   maxWidth: "600px",
+    // },
+    maxWidth: "80%",
+    margin: "auto",
+    cursor: "pointer",
+  },
+  agenda: {
+    padding: "4% 4%",
+  },
+}));
+
+const AgendaView: React.FC<{ data?: ICardInfo }> = ({ data }) => {
+  const { contentTransMode } = useContext(SettingContext)!;
+  const classes = useStyles();
+  const { path } = useRouteMatch();
+  const getCharaName = useCharaName(contentTransMode);
+
+  if (!data) {
+    // loading
+    return (
+      <Box className={classes.agendaWrapper}>
+        <Paper className={classes.agenda}>
+          <Grid
+            container
+            alignItems="center"
+            spacing={2}
+            justify="space-between"
+          >
+            <Grid
+              item
+              xs={5}
+              md={4}
+              container
+              direction="row"
+              spacing={1}
+              justify="center"
+            >
+              <Grid item xs={12} md={6}>
+                <CardThumbSkeleton></CardThumbSkeleton>
+              </Grid>
+              <Grid item xs={12} md={6}>
+                <CardThumbSkeleton></CardThumbSkeleton>
+              </Grid>
+            </Grid>
+            <Grid item xs={6} md={7}>
+              <Typography variant="body1">
+                <Skeleton variant="text" width="70%"></Skeleton>
+              </Typography>
+              <Typography variant="body2">
+                <Skeleton variant="text" width="30%"></Skeleton>
+              </Typography>
+            </Grid>
+          </Grid>
+        </Paper>
+      </Box>
+    );
+  }
+  return (
+    <Link
+      to={path + "/" + data.id}
+      className={classes.agendaWrapper}
+      style={{ textDecoration: "none" }}
+    >
+      <Paper className={classes.agenda}>
+        <Grid container alignItems="center" spacing={2} justify="space-between">
+          <Grid
+            item
+            xs={5}
+            md={4}
+            container
+            direction="row"
+            spacing={1}
+            justify="center"
+          >
+            <Grid item xs={12} md={6}>
+              <CardThumb cardId={data.id} />
+            </Grid>
+            {data.rarity >= 3 ? (
+              <Grid item xs={12} md={6}>
+                <CardThumb cardId={data.id} trained />
+              </Grid>
+            ) : null}
+          </Grid>
+          <Grid item xs={6} md={7}>
+            <Grid container direction="column" spacing={1}>
+              <Grid item>
+                <ContentTrans
+                  mode={contentTransMode}
+                  contentKey={`card_prefix:${data.id}`}
+                  original={data.prefix}
+                  originalProps={{ variant: "body1" }}
+                  translatedProps={{ variant: "body1" }}
+                />
+              </Grid>
+              <Grid item>
+                <Typography variant="body2" color="textSecondary">
+                  {getCharaName(data.characterId)}
+                </Typography>
+              </Grid>
+            </Grid>
+          </Grid>
+        </Grid>
+      </Paper>
+    </Link>
+  );
+};
+
+export default AgendaView;
